+++ src/main/java/nextstep/subway/application/dto/LineResponse.java
@@ -1,16 +1,18 @@
-package nextstep.subway.applicaion.dto;
+package nextstep.subway.application.dto;
 
 import java.time.LocalDateTime;
 
-public class StationResponse {
+public class LineResponse {
     private Long id;
     private String name;
+    private String color;
     private LocalDateTime createdDate;
     private LocalDateTime modifiedDate;
 
-    public StationResponse(Long id, String name, LocalDateTime createdDate, LocalDateTime modifiedDate) {
+    public LineResponse(Long id, String name, String color, LocalDateTime createdDate, LocalDateTime modifiedDate) {
         this.id = id;
         this.name = name;
+        this.color = color;
         this.createdDate = createdDate;
         this.modifiedDate = modifiedDate;
     }
@@ -21,6 +23,10 @@ public class StationResponse {
 
     public String getName() {
         return name;
+    }
+
+    public String getColor() {
+        return color;
     }
 
     public LocalDateTime getCreatedDate() {
