+++ src/main/java/nextstep/subway/application/dto/StationResponse.java
@@ -1,18 +1,16 @@
-package nextstep.subway.applicaion.dto;
+package nextstep.subway.application.dto;
 
 import java.time.LocalDateTime;
 
-public class LineResponse {
+public class StationResponse {
     private Long id;
     private String name;
-    private String color;
     private LocalDateTime createdDate;
     private LocalDateTime modifiedDate;
 
-    public LineResponse(Long id, String name, String color, LocalDateTime createdDate, LocalDateTime modifiedDate) {
+    public StationResponse(Long id, String name, LocalDateTime createdDate, LocalDateTime modifiedDate) {
         this.id = id;
         this.name = name;
-        this.color = color;
         this.createdDate = createdDate;
         this.modifiedDate = modifiedDate;
     }
@@ -23,10 +21,6 @@ public class LineResponse {
 
     public String getName() {
         return name;
-    }
-
-    public String getColor() {
-        return color;
     }
 
     public LocalDateTime getCreatedDate() {
